fix(ss-2666): assertUnreachable moved to helpers
