--- src/libs/assertUnreachable.ts
@@ -1,3 +1,0 @@
-export function assertUnreachable(_x: never): never {
-  throw new Error('Didn\'t expect to get here');
-}
